Regression in https://github.com/libretro/RetroArch/commit/9c73d3305dbb0a81e14d21e13aac65444fb3bf2d - joykey needs to be uint64_t and then is later cast to uint16_t - is apparently important for hotkeys

+++ input/drivers/android_input.c
@@ -1361,7 +1361,7 @@ static void android_input_poll_memcpy(android_input_t *android)
 
 static bool android_input_key_pressed(android_input_t *android, int key)
 {
-   uint16_t joykey;
+   uint64_t joykey;
    uint32_t joyaxis;
    rarch_joypad_info_t joypad_info;
    joypad_info.joy_idx        = 0;
@@ -1383,7 +1383,7 @@ static bool android_input_key_pressed(android_input_t *android, int key)
       ? input_config_binds[0][key].joyaxis 
       : joypad_info.auto_binds[key].joyaxis;
 
-   if (joykey != NO_BTN && android->joypad->button(joypad_info.joy_idx, joykey))
+   if ((uint16_t)joykey != NO_BTN && android->joypad->button(joypad_info.joy_idx, (uint16_t)joykey))
       return true;
    else if (((float)abs(android->joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
       return true;
@@ -1480,12 +1480,12 @@ static int16_t android_input_state(void *data,
             {
                bool res               = false;
                /* Auto-binds are per joypad, not per user. */
-               const uint16_t joykey  = (binds[port][i].joykey != NO_BTN)
+               const uint64_t joykey  = (binds[port][i].joykey != NO_BTN)
                   ? binds[port][i].joykey : joypad_info.auto_binds[i].joykey;
                const uint32_t joyaxis = (binds[port][i].joyaxis != AXIS_NONE)
                   ? binds[port][i].joyaxis : joypad_info.auto_binds[i].joyaxis;
-               if (joykey != NO_BTN && android->joypad->button(
-                        joypad_info.joy_idx, joykey))
+               if ((uint16_t)joykey != NO_BTN && android->joypad->button(
+                        joypad_info.joy_idx, (uint16_t)joykey))
                   res = true;
                else if (((float)abs(android->joypad->axis(
                               joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
@@ -1500,12 +1500,12 @@ static int16_t android_input_state(void *data,
          {
             bool res               = false;
             /* Auto-binds are per joypad, not per user. */
-            const uint16_t joykey  = (binds[port][id].joykey != NO_BTN)
+            const uint64_t joykey  = (binds[port][id].joykey != NO_BTN)
                ? binds[port][id].joykey : joypad_info.auto_binds[id].joykey;
             const uint32_t joyaxis = (binds[port][id].joyaxis != AXIS_NONE)
                ? binds[port][id].joyaxis : joypad_info.auto_binds[id].joyaxis;
-            if (joykey != NO_BTN && android->joypad->button(
-                     joypad_info.joy_idx, joykey))
+            if ((uint16_t)joykey != NO_BTN && android->joypad->button(
+                     joypad_info.joy_idx, (uint16_t)joykey))
                res = true;
             else if (((float)abs(android->joypad->axis(
                            joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
