+++ input/drivers/rwebinput_input.c
@@ -487,7 +487,7 @@ static bool rwebinput_is_pressed(rwebinput_input_t *rwebinput,
       if (bind->valid)
       {
          /* Auto-binds are per joypad, not per user. */
-         const uint16_t joykey  = (binds[id].joykey != NO_BTN)
+         const uint64_t joykey  = (binds[id].joykey != NO_BTN)
             ? binds[id].joykey : joypad_info.auto_binds[id].joykey;
          const uint32_t joyaxis = (binds[id].joyaxis != AXIS_NONE)
             ? binds[id].joyaxis : joypad_info.auto_binds[id].joyaxis;
@@ -495,7 +495,7 @@ static bool rwebinput_is_pressed(rwebinput_input_t *rwebinput,
          if (port == 0 && !!rwebinput_mouse_state(&rwebinput->mouse,
                   bind->mbutton, false))
             return true;
-         if (joykey != NO_BTN && rwebinput->joypad->button(joypad_info.joy_idx, joykey))
+         if ((uint16_t)joykey != NO_BTN && rwebinput->joypad->button(joypad_info.joy_idx, (uint16_t)joykey))
             return true;
          if (((float)abs(rwebinput->joypad->axis(joypad_info.joy_idx, joyaxis)) / 0x8000) > joypad_info.axis_threshold)
             return true;
